requirements.txt
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
Django==2.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.17.0
idna==2.8
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.15.2
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
prompt-toolkit==3.0.2
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
PyJWT==1.7.1
python-dotenv==0.10.5
pytz==2019.3
razorpay==1.2.0
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
